SITL: support 32 bit system IDs
The ADSB, SoloGimbal, MAVLink gimbal v2 and camera v2 simulators
truncated the vehicle sysid they learn from its heartbeat, matched
commands against the payload target byte (which is zero when the
target travels in the extended header), and addressed replies through
payload structs. The simulated heartbeats are also serialised properly
rather than writing the mavlink_message_t memory as wire bytes, which
is no longer the wire layout now that sysid is 32 bit.

+++ libraries/SITL/SIM_ADSB.h
@@ -76,7 +76,7 @@ private:
     
     uint32_t last_heartbeat_ms;
     bool seen_heartbeat = false;
-    uint8_t vehicle_system_id;
+    uint32_t vehicle_system_id;
     uint8_t vehicle_component_id;
 
     struct {
